requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
django-registration==3.0.1
confusable-homoglyphs==3.2.0
django-friendly-tag-loader==1.3.1
django-widget-tweaks==1.4.3
django==2.0.13
djangorestframework==3.9.1
fasteners==0.14.1
monotonic==1.5            # via fasteners
pillow==5.4.1
psycopg2-binary==2.7.7
pytz==2018.9              # via django
six==1.12.0               # via fasteners
gunicorn==19.9.0
numpy==1.17.4
opencv-python==4.1.2.30
openslide-python==1.1.1
uWSGI==2.0.18
